Final
I could not find the way to save the buildings in the grid. I tried everthing but in the grid system I wrote every grid square is working from the same script. I could not find a way to locate the spesific location of every grid because they all seen in the same place.